move things db_utils -> db
