+++ apps/wizard/app_pages/indicator_upgrade/utils.py
@@ -8,7 +8,9 @@ from rapidfuzz import fuzz
 from structlog import get_logger
 
 from apps.utils.map_datasets import get_grapher_changes
-from etl.db_utils import config, get_all_datasets, get_connection, get_dataset_charts, get_variables_in_dataset
+from etl import config
+from etl.db import get_all_datasets, get_dataset_charts, get_variables_in_dataset
+from etl.db_utils import get_connection
 from etl.git_helpers import get_changed_files
 from etl.match_variables import find_mapping_suggestions, preliminary_mapping
 from etl.version_tracker import VersionTracker
