+++ etl/db.py
@@ -1,11 +1,17 @@
+import warnings
 from typing import Any, Dict, List, Optional, cast
 
 import pandas as pd
-import streamlit as st
+import pymysql
+import structlog
+import validators
 from sqlalchemy.orm import Session
 
 from etl.config import OWID_ENV, OWIDEnv
+from etl.db_utils import get_connection
 from etl.grapher_model import Variable
+
+log = structlog.get_logger()
 
 
 def load_variables_in_dataset(dataset_uris: List[str]) -> List[Variable]:
@@ -105,31 +111,36 @@ def ensure_variable(
     return variable
 
 
-def variable_data_df_from_s3(
-    engine: Engine,
-    variable_ids: List[int] = [],
-    workers: int = 1,
-    value_as_str: bool = True,
-) -> pd.DataFrame:
-    """Fetch data from S3 and add entity code and name from DB."""
-    with concurrent.futures.ThreadPoolExecutor(max_workers=workers) as executor:
-        results = list(executor.map(_fetch_data_df_from_s3, variable_ids))
+# def variable_data_df_from_s3(
+#     engine: Engine,
+#     variable_ids: List[int] = [],
+#     workers: int = 1,
+#     value_as_str: bool = True,
+# ) -> pd.DataFrame:
+#     """Fetch data from S3 and add entity code and name from DB."""
+#     with concurrent.futures.ThreadPoolExecutor(max_workers=workers) as executor:
+#         results = list(executor.map(_fetch_data_df_from_s3, variable_ids))
 
-    if isinstance(results, list) and all(isinstance(df, pd.DataFrame) for df in results):
-        df = pd.concat(cast(List[pd.DataFrame], results))
-    else:
-        raise TypeError(f"results must be a list of pd.DataFrame, got {type(results)}")
+#     if isinstance(results, list) and all(isinstance(df, pd.DataFrame) for df in results):
+#         df = pd.concat(cast(List[pd.DataFrame], results))
+#     else:
+#         raise TypeError(f"results must be a list of pd.DataFrame, got {type(results)}")
 
-    # we work with strings and convert to specific types later
-    if value_as_str:
-        df["value"] = df["value"].astype("string")
+#     # we work with strings and convert to specific types later
+#     if value_as_str:
+#         df["value"] = df["value"].astype("string")
 
-    with Session(engine) as session:
-        res = add_entity_code_and_name(session, df)
-        return res
+#     with Session(engine) as session:
+#         res = add_entity_code_and_name(session, df)
+#         return res
 
 
-#################### OLD
+##############################################################################################
+# TO BE REVIEWED:
+# This is code that could be deprecated / removed?
+##############################################################################################
+
+
 def get_dataset_id(
     dataset_name: str, db_conn: Optional[pymysql.Connection] = None, version: Optional[str] = None
 ) -> Any:
@@ -212,4 +223,352 @@ def get_variables_in_dataset(
     with warnings.catch_warnings():
         warnings.simplefilter("ignore", UserWarning)
         variables_data = pd.read_sql(query, con=db_conn)
+    return variables_data
+
+
+def get_all_datasets(archived: bool = True, db_conn: Optional[pymysql.Connection] = None) -> pd.DataFrame:
+    """Get all datasets in database.
+
+    Parameters
+    ----------
+    db_conn : pymysql.connections.Connection
+        Connection to database. Defaults to None, in which case a default connection is created (uses etl.config).
+
+    Returns
+    -------
+    datasets : pd.DataFrame
+        All datasets in database. Table with three columns: dataset ID, dataset name, dataset namespace.
+    """
+    if db_conn is None:
+        db_conn = get_connection()
+
+    query = " SELECT namespace, name, id, updatedAt, isArchived FROM datasets"
+    if not archived:
+        query += " WHERE isArchived = 0"
+    datasets = pd.read_sql(query, con=db_conn)
+    return datasets.sort_values(["name", "namespace"])
+
+
+def get_info_for_etl_datasets(db_conn: Optional[pymysql.Connection] = None) -> pd.DataFrame:
+    if db_conn is None:
+        db_conn = get_connection()
+
+    # First, increase the GROUP_CONCAT limit, to avoid the list of chart ids to be truncated.
+    GROUP_CONCAT_MAX_LEN = 4096
+    cursor = db_conn.cursor()
+    cursor.execute(f"SET SESSION group_concat_max_len = {GROUP_CONCAT_MAX_LEN};")
+    db_conn.commit()
+
+    query = """\
+    SELECT
+        q1.datasetId AS dataset_id,
+        d.name AS dataset_name,
+        q1.etlPath AS etl_path,
+        d.isArchived AS is_archived,
+        d.isPrivate AS is_private,
+        q2.chartIds AS chart_ids,
+        q2.updatePeriodDays AS update_period_days
+    FROM
+        (SELECT
+            datasetId,
+            MIN(catalogPath) AS etlPath
+        FROM
+            variables
+        WHERE
+            catalogPath IS NOT NULL
+        GROUP BY
+            datasetId) q1
+    LEFT JOIN
+        (SELECT
+            d.id AS datasetId,
+            d.isArchived,
+            d.isPrivate,
+            d.updatePeriodDays,
+            GROUP_CONCAT(DISTINCT c.id) AS chartIds
+        FROM
+            datasets d
+            JOIN variables v ON v.datasetId = d.id
+            JOIN chart_dimensions cd ON cd.variableId = v.id
+            JOIN charts c ON c.id = cd.chartId
+            JOIN chart_configs cc ON c.configId = cc.id
+        WHERE
+            json_extract(cc.full, "$.isPublished") = TRUE
+        GROUP BY
+            d.id) q2
+        ON q1.datasetId = q2.datasetId
+    JOIN
+        datasets d ON q1.datasetId = d.id
+    ORDER BY
+        q1.datasetId ASC;
+
+    """
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", UserWarning)
+        df = pd.read_sql(query, con=db_conn)
+
+    if max([len(row) for row in df["chart_ids"] if row is not None]) == GROUP_CONCAT_MAX_LEN:
+        log.error(
+            f"The value of group_concat_max_len (set to {GROUP_CONCAT_MAX_LEN}) has been exceeded."
+            "This means that the list of chart ids will be incomplete in some cases. Consider increasing it."
+        )
+
+    # Get mapping of chart ids to slugs.
+    chart_id_to_slug = get_charts_slugs(db_conn=db_conn).set_index("chart_id")["chart_slug"].to_dict()
+
+    # Instead of having a string of chart ids, make chart_ids a column with lists of integers.
+    df["chart_ids"] = [
+        [int(chart_id) for chart_id in chart_ids.split(",")] if chart_ids else [] for chart_ids in df["chart_ids"]
+    ]
+    # Add a column with lists of chart slugs.
+    # For each row, it will be a list of tuples (chart_id, chart_slug),
+    # e.g. [(123, "chart-slug"), (234, "another-chart-slug"), ...].
+    df["chart_slugs"] = [
+        [(chart_id, chart_id_to_slug[chart_id]) for chart_id in chart_ids] if chart_ids else []
+        for chart_ids in df["chart_ids"]
+    ]
+
+    # Add chart analytics.
+    views_df = get_charts_views(db_conn=db_conn).set_index("slug")
+    # Create a column for each of the views metrics.
+    # For each row, it will be a list of tuples (chart_id, views),
+    # e.g. [(123, 1000), (234, 2000), ...].
+    for metric in views_df.columns:
+        df[metric] = [
+            [
+                (chart_id, views_df[metric][chart_id_to_slug[chart_id]])
+                for chart_id in chart_ids
+                if chart_id_to_slug[chart_id] in views_df.index
+            ]
+            if chart_ids
+            else []
+            for chart_ids in df["chart_ids"]
+        ]
+
+    # Make is_archived and is_private boolean columns.
+    df["is_archived"] = df["is_archived"].astype(bool)
+    df["is_private"] = df["is_private"].astype(bool)
+
+    # Sanity check.
+    unknown_channels = set([etl_path.split("/")[0] for etl_path in set(df["etl_path"])]) - {"grapher"}
+    if len(unknown_channels) > 0:
+        log.error(
+            "Variables in grapher DB are expected to come only from ETL grapher channel, "
+            f"but other channels were found: {unknown_channels}"
+        )
+
+    # Create a column with the step name.
+    # First assume all steps are public (hence starting with "data://").
+    # Then edit private steps so they start with "data-private://".
+    df["step"] = ["data://" + "/".join(etl_path.split("#")[0].split("/")[:-1]) for etl_path in df["etl_path"]]
+    df.loc[df["is_private"], "step"] = df[df["is_private"]]["step"].str.replace("data://", "data-private://")
+
+    return df
+
+
+def get_charts_slugs(db_conn: Optional[pymysql.Connection] = None) -> pd.DataFrame:
+    if db_conn is None:
+        db_conn = get_connection()
+
+    # Get a dataframe chart_id,char_slug, for all charts that have variables with an ETL path.
+    query = """\
+    SELECT
+        c.id AS chart_id,
+        cc.slug AS chart_slug
+    FROM charts c
+    JOIN chart_configs cc ON c.configId = cc.id
+    LEFT JOIN chart_dimensions cd ON c.id = cd.chartId
+    LEFT JOIN variables v ON cd.variableId = v.id
+    WHERE
+        v.catalogPath IS NOT NULL
+    ORDER BY
+        c.id ASC;
+    """
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", UserWarning)
+        df = pd.read_sql(query, con=db_conn)
+
+    # Remove duplicated rows.
+    df = df.drop_duplicates().reset_index(drop=True)
+
+    if len(df[df.duplicated(subset="chart_id")]) > 0:
+        log.warning("There are duplicated chart ids in the chart_ids and slugs table.")
+
+    return df
+
+
+def get_charts_views(db_conn: Optional[pymysql.Connection] = None) -> pd.DataFrame:
+    if db_conn is None:
+        db_conn = get_connection()
+
+    # Assumed base url for all charts.
+    base_url = "https://ourworldindata.org/grapher/"
+
+    # Note that for now we extract data for all dates.
+    # It seems that the table only has data for the last day.
+    query = f"""\
+    SELECT
+        url,
+        views_7d,
+        views_14d,
+        views_365d
+    FROM
+        analytics_pageviews
+    WHERE
+        url LIKE '{base_url}%';
+    """
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", UserWarning)
+        df = pd.read_sql(query, con=db_conn)
+
+    # For some reason, there are spurious urls, clean some of them.
+    # Note that validators.url() returns a ValidationError object (instead of False) when the url has spaces.
+    is_url_invalid = [(validators.url(url) is False) or (" " in url) for url in df["url"]]
+    df = df.drop(df[is_url_invalid].index).reset_index(drop=True)
+
+    # Note that some of the returned urls may still be invalid, for example "https://ourworldindata.org/grapher/132".
+
+    # Add chart slug.
+    df["slug"] = [url.replace(base_url, "") for url in df["url"]]
+
+    # Remove url.
+    df = df.drop(columns=["url"], errors="raise")
+
+    if len(df[df.duplicated(subset="slug")]) > 0:
+        log.warning("There are duplicated slugs in the chart analytics table.")
+
+    return df
+
+
+def get_dataset_charts(dataset_ids: List[str], db_conn: Optional[pymysql.Connection] = None) -> pd.DataFrame:
+    if db_conn is None:
+        db_conn = get_connection()
+
+    dataset_ids_str = ", ".join(map(str, dataset_ids))
+
+    query = f"""
+    SELECT
+        d.id AS dataset_id,
+        d.name AS dataset_name,
+        q2.chartIds AS chart_ids
+    FROM
+        (SELECT
+            d.id,
+            d.name
+        FROM
+            datasets d
+        WHERE
+            d.id IN ({dataset_ids_str})) d
+    LEFT JOIN
+        (SELECT
+            v.datasetId,
+            GROUP_CONCAT(DISTINCT c.id) AS chartIds
+        FROM
+            variables v
+            JOIN chart_dimensions cd ON cd.variableId = v.id
+            JOIN charts c ON c.id = cd.chartId
+        WHERE
+            v.datasetId IN ({dataset_ids_str})
+        GROUP BY
+            v.datasetId) q2
+        ON d.id = q2.datasetId
+    ORDER BY
+        d.id ASC;
+    """
+
+    # First, increase the GROUP_CONCAT limit, to avoid the list of chart ids to be truncated.
+    with db_conn.cursor() as cursor:
+        cursor.execute("SET SESSION group_concat_max_len = 10000;")
+
+    if len(dataset_ids) == 0:
+        return pd.DataFrame({"dataset_id": [], "dataset_name": [], "chart_ids": []})
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", UserWarning)
+        df = pd.read_sql(query, con=db_conn)
+
+    # Instead of having a string of chart ids, make chart_ids a column with lists of integers.
+    df["chart_ids"] = [
+        [int(chart_id) for chart_id in chart_ids.split(",")] if chart_ids else [] for chart_ids in df["chart_ids"]
+    ]
+
+    return df
+
+
+def get_variables_data(
+    filter: Optional[Dict[str, Any]] = None,
+    condition: Optional[str] = "OR",
+    db_conn: Optional[pymysql.Connection] = None,
+) -> pd.DataFrame:
+    """Get data from variables table, given a certain condition.
+
+    Parameters
+    ----------
+    filter : Optional[Dict[str, Any]], optional
+        Filter to apply to the data, which must contain a field name and a list of field values,
+        e.g. {"id": [123456, 234567, 345678]}.
+        In principle, multiple filters can be given.
+    condition : Optional[str], optional
+        In case multiple filters are given, this parameter specifies whether the output filters should be the union
+        ("OR") or the intersection ("AND").
+    db_conn : pymysql.Connection
+        Connection to database. Defaults to None, in which case a default connection is created (uses etl.config).
+
+    Returns
+    -------
+    df : pd.DataFrame
+        Variables data.
+
+    """
+    # NOTE: This function should be optimized. Instead of fetching data for each filter, their conditions should be
+    # combined with OR or AND before executing the query.
+
+    # Initialize an empty dataframe.
+    if filter is not None:
+        df = pd.DataFrame({"id": []}).astype({"id": int})
+        for field_name, field_values in filter.items():
+            _df = _get_variables_data_with_filter(field_name=field_name, field_values=field_values, db_conn=db_conn)
+            if condition == "OR":
+                df = pd.concat([df, _df], axis=0)
+            elif condition == "AND":
+                df = pd.merge(df, _df, on="id", how="inner")
+            else:
+                raise ValueError(f"Invalid condition: {condition}")
+    else:
+        # Fetch data for all variables.
+        df = _get_variables_data_with_filter(db_conn=db_conn)
+
+    return df
+
+
+def _get_variables_data_with_filter(
+    field_name: Optional[str] = None,
+    field_values: Optional[List[Any]] = None,
+    db_conn: Optional[pymysql.Connection] = None,
+) -> Any:
+    if db_conn is None:
+        db_conn = get_connection()
+
+    if field_values is None:
+        field_values = []
+
+    # Construct the SQL query with a placeholder for each value in the list.
+    query = "SELECT * FROM variables"
+
+    if (field_name is not None) and (len(field_values) > 0):
+        query += f"\nWHERE {field_name} IN ({', '.join(['%s'] * len(field_values))});"
+
+    # Execute the query.
+    with warnings.catch_warnings():
+        warnings.simplefilter("ignore", UserWarning)
+        variables_data = pd.read_sql(query, con=db_conn, params=field_values)
+
+    assert set(variables_data[field_name]) <= set(field_values), f"Unexpected values for {field_name}."
+
+    # Warn about values that were not found.
+    missing_values = set(field_values) - set(variables_data[field_name])
+    if len(missing_values) > 0:
+        log.warning(f"Values of {field_name} not found in database: {missing_values}")
+
     return variables_data
