+++ etl/explorer_helpers.py
@@ -5,7 +5,7 @@ import pandas as pd
 from structlog import get_logger
 
 from etl import config
-from etl.db_utils import get_variables_data
+from etl.db import get_variables_data
 from etl.files import upload_file_to_server
 from etl.paths import EXPLORERS_DIR
