+++ etl/match_variables.py
@@ -9,7 +9,8 @@ import rich_click as click
 from rapidfuzz import fuzz
 from structlog import get_logger
 
-from etl import db_utils
+from etl.db import get_dataset_id, get_variables_in_dataset
+from etl.db_utils import get_connection
 
 # If True, identical variables will be matched automatically (by string comparison).
 # If False, variables with identical names will appear in comparison.
@@ -124,19 +125,15 @@ def main(
     if Path(output_file).suffix != ".json":
         raise ValueError(f"`output_file` ({output_file}) should point to a JSON file ('*.json')!")
 
-    with db_utils.get_connection() as db_conn:
+    with get_connection() as db_conn:
         # Get old and new dataset ids.
-        old_dataset_id = db_utils.get_dataset_id(db_conn=db_conn, dataset_name=old_dataset_name)
-        new_dataset_id = db_utils.get_dataset_id(db_conn=db_conn, dataset_name=new_dataset_name)
+        old_dataset_id = get_dataset_id(db_conn=db_conn, dataset_name=old_dataset_name)
+        new_dataset_id = get_dataset_id(db_conn=db_conn, dataset_name=new_dataset_name)
 
         # Get variables from old dataset that have been used in at least one chart.
-        old_indicators = db_utils.get_variables_in_dataset(
-            db_conn=db_conn, dataset_id=old_dataset_id, only_used_in_charts=True
-        )
+        old_indicators = get_variables_in_dataset(db_conn=db_conn, dataset_id=old_dataset_id, only_used_in_charts=True)
         # Get all variables from new dataset.
-        new_indicators = db_utils.get_variables_in_dataset(
-            db_conn=db_conn, dataset_id=new_dataset_id, only_used_in_charts=False
-        )
+        new_indicators = get_variables_in_dataset(db_conn=db_conn, dataset_id=new_dataset_id, only_used_in_charts=False)
 
     # Manually map old variable names to new variable names.
     mapping = map_old_and_new_indicators(
